fix(lap-validity): drop useless i32::from conversion (clippy)

+++ sidecar/src/lap_validity.rs
@@ -108,7 +108,7 @@ impl LapValidityDetector {
             existing
         } else {
             let created =
-                storage.start_session(Some(i32::from(packet.sled.car_ordinal)), sidecar_version)?;
+                storage.start_session(Some(packet.sled.car_ordinal), sidecar_version)?;
             self.current_session_id = Some(created);
             created
         };
@@ -165,7 +165,7 @@ impl LapValidityDetector {
 
         storage.end_session(prior_session_id)?;
         let new_session_id =
-            storage.start_session(Some(i32::from(packet.sled.car_ordinal)), sidecar_version)?;
+            storage.start_session(Some(packet.sled.car_ordinal), sidecar_version)?;
         storage.ensure_lap(new_session_id, packet.lap_number, packet.sled.timestamp_ms)?;
 
         self.current_session_id = Some(new_session_id);
